Psudo Restructure Code and Rework Timing
The Restructure allows for the system to call the same functions for
sending and receiving for either the Daughter or Master, to be able to
symmetric for making testing easier and for meeting specs/reqs. This
also allows the system to have comparisons for different types of tests
instead of commenting out things when not using them for specific tests,
instead changing a condition or a number. I reworked the timing to now
be based on the computer's date and time when sending and receiving
instead of starting and stopping timers since i couldn't get that
figured out.